requirements.txt
aiomysql==0.2.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
fastapi
greenlet==3.2.2
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
pydantic
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.1.0
requests
selenium==4.15.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.46.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn
websocket-client==1.8.0
wsproto==1.2.0
python-multipart
numpy
Pillow
scipy
opencv-python
torch==2.3.0
torchvision==0.18.0
ultralytics>=8.3.153
python-multipart
pydantic
requests
matplotlib
scikit-learn
easyocr
